+++ tests/testthat/test-model_parameters.aov.R
@@ -1,6 +1,6 @@
 .runThisTest <- Sys.getenv("RunAllparametersTests") == "yes"
 
-if (.runThisTest || Sys.getenv("USER") == "travis") {
+if (.runThisTest) {
   if (require("insight") && require("testthat") && require("lme4") && require("parameters")) {
 
     if (requireNamespace("effectsize")) {
@@ -14,7 +14,7 @@ if (.runThisTest || Sys.getenv("USER") == "travis") {
         mp <- model_parameters(model, omega_squared = "partial", eta_squared = "partial", epsilon_squared = TRUE)
         testthat::expect_equal(sum(mp$df), 149)
         testthat::expect_equal(colnames(mp), c("Parameter", "Sum_Squares", "df", "Mean_Square", "F", "p",
-                                               "Omega2_partial", "Eta2_partial", "Epsilon2"))
+                                               "Omega2_partial", "Eta2_partial", "Epsilon2_partial"))
 
         model <- aov(Sepal.Length ~ Species * Cat1 * Cat2, data = iris)
         testthat::expect_equal(sum(model_parameters(model, omega_squared = "raw", eta_squared = "partial", epsilon_squared = TRUE)$df), 149)