+++ NEWS.md
@@ -1,5 +1,9 @@
 # parameters 0.9.1
 
+## Breaking changes
+
+* The default-method for effect sizes in `model_parameters()` for Anova-models (i.e. when arguments `omega_squared`, `eta_squared` or `epsilon_squared` are set to `TRUE` is now `"partial"`, as initially intended.
+
 ## New supported model classes
 
 * `model_parameters()` supports `Gam` models (*gam*).
@@ -7,6 +11,7 @@
 ## Changes to functions
 
 * `describe_distribution()` now includes the name of the centrality index in the `CI`-column, when `centrality = "all"`.
+* `model_parameters()` for Anova-models (of class `aov`, `anova` etc.) gains a `ci`-argument, to add confidence intervals to effect size parameters.
 
 # parameters 0.9.0